move zcsr_kron
