+++ qutip/cy/setup.py
@@ -8,7 +8,8 @@ import os
 
 dir_path = os.path.dirname(os.path.realpath(__file__))
 
-exts = ['spmatfuncs', 'stochastic', 'sparse_utils', 'graph_utils', 'interpolate']
+exts = ['spmatfuncs', 'stochastic', 'sparse_utils', 'graph_utils', 'interpolate',
+        'spmath']
 
 _compiler_flags = ['-w', '-ffast-math', '-O3', '-march=native', '-funroll-loops']
